+++ extras/qpsclient.py
@@ -35,6 +35,10 @@ class QPSSpider(Spider):
             self.download_delay = float(self.download_delay)
 
     async def yield_seeds(self):
+        for seed in self.start_requests():
+            yield seed
+
+    def start_requests(self):
         url = self.benchurl
         if self.latency is not None:
             url += f"?latency={self.latency}"