+++ scrapy/spiders/init.py
@@ -30,7 +30,7 @@ class InitSpider(Spider):
         )
 
     async def yield_seeds(self) -> AsyncIterator[Any]:
-        async for seed in super().yield_seeds():
+        for seed in self.start_requests():
             yield seed
 
     def start_requests(self) -> Iterable[Request]: