+++ tests/CrawlerProcess/asyncio_enabled_reactor.py
@@ -9,7 +9,8 @@ class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
     async def yield_seeds(self):
-        return []
+        return
+        yield
 
 
 process = CrawlerProcess(