+++ tests/CrawlerProcess/asyncio_enabled_reactor_different_loop.py
@@ -16,7 +16,8 @@ class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
     async def yield_seeds(self):
-        return []
+        return
+        yield
 
 
 process = CrawlerProcess(