+++ tests/CrawlerProcess/asyncio_enabled_reactor_same_loop.py
@@ -17,7 +17,8 @@ class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
     async def yield_seeds(self):
-        return []
+        return
+        yield
 
 
 process = CrawlerProcess(