+++ tests/CrawlerProcess/reactor_select.py
@@ -11,7 +11,8 @@ class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
     async def yield_seeds(self):
-        return []
+        return
+        yield
 
 
 process = CrawlerProcess(settings={})