+++ tests/test_crawler.py
@@ -581,7 +581,8 @@ class NoRequestsSpider(scrapy.Spider):
     name = "no_request"
 
     async def yield_seeds(self):
-        return []
+        return
+        yield
 
 
 @pytest.mark.usefixtures("reactor_pytest")